backend/requirements.txt
flask
PyJWT[crypto]
yfinance
pandas
numpy
scikit-learn
python-dotenv
gunicorn
